Rename Zombie.h/cpp to Aumtlich.h/cpp and update includes.
